Decompose Documents.tsx into thinner composer + sibling files
frontend/src/views/Documents.tsx had grown to 1,725 lines. Five
extractions land it at 915 lines (-47%) without changing public API,
DOM, or behavior:

- Documents.styles.ts: all 50 styled-components as named exports.
  Cross-component selectors (DocumentCardWrapper:hover &, ListItem:hover &)
  require same-module scope, so all styles stay in one file. Renamed
  three styled-components to free names for the new React components:
  DocumentsListContainer (outer section wrapper) -> DocumentsSection,
  DocumentsListView (inner list-table) -> DocumentsListContainer,
  DocumentsCompactView (inner compact list) -> DocumentsCompactContainer.
  Dropped the unused MetaSeparator styled-component (dead code).
- DocumentsGridView.tsx: the grid-view branch (cards with checkbox /
  preview / body / footer).
- DocumentsListView.tsx: the list-view branch including the table header
  with select-all checkbox.
- DocumentsCompactView.tsx: the compact-view branch.
- documentsUtils.ts: getDocumentType(doc) helper shared by the grid and
  list views.

The hero section, advanced-filters popup, stats grid, section header
with view-mode toggle, fetch-more sentinel, empty-state branches,
context menu, modals, and the DocumentIcon SVG fallback stay inline.
All 11 Apollo reactive vars stay at the composer; props are drilled
down to the three view components (children stay pure-presentational,
matching the ExtractsTabContent.tsx precedent). Public API (Documents
named export) is unchanged; importers in App.tsx, Documents.ct.tsx,
and DocumentsTestWrapper.tsx need no changes.

Verified: tsc --noEmit clean; prettier clean; all 10 component tests
in Documents.ct.tsx pass; all 4 regression cases in
Documents.refetch-shape.test.ts pass.

Final slice of the multi-PR decomposition tracked by #1446 (after
#1459 / ModernDocumentItem, #1578 / Corpuses, #1586 / ChatMessage,
#1580 / DocumentKnowledgeBase partial, #1639 / ChatTray).

Closes #1446
