+++ frontend/src/views/Documents.tsx
@@ -8,11 +8,7 @@ import {
 import { useNavigate } from "react-router-dom";
 import { toast } from "react-toastify";
 import _ from "lodash";
-import styled from "styled-components";
-import {
-  OS_LEGAL_COLORS,
-  accentAlpha,
-} from "../assets/configurations/osLegalStyles";
+import { OS_LEGAL_COLORS } from "../assets/configurations/osLegalStyles";
 import {
   PageContainer,
   ContentContainer,
@@ -31,8 +27,6 @@ import {
   StatGrid,
   Button,
   EmptyState,
-  Chip,
-  Avatar,
 } from "@os-legal/ui";
 import type { FilterTabItem } from "@os-legal/ui";
 import {
@@ -40,9 +34,6 @@ import {
   Grid,
   List,
   AlignJustify,
-  MoreVertical,
-  FileText,
-  Loader2,
   SlidersHorizontal,
   X,
   AlertCircle,
@@ -99,8 +90,6 @@ import { FetchMoreOnVisible } from "../components/widgets/infinite_scroll/FetchM
 import { FetchMoreFooter } from "../components/widgets/infinite_scroll/FetchMoreFooter";
 import { LoadingOverlay } from "../components/common/LoadingOverlay";
 import { navigateToDocument } from "../utils/navigationUtils";
-import { formatFileSize, formatRelativeTime } from "../utils/formatters";
-import { getCreatorInitials } from "../utils/userDisplay";
 import {
   VIEW_MODES,
   STATUS_FILTERS,
@@ -108,6 +97,26 @@ import {
   type ViewMode,
   type StatusFilter,
 } from "../assets/configurations/constants";
+import {
+  ActionButtons,
+  ClearFiltersButton,
+  DocumentsSection,
+  FilterBadge,
+  FilterButton,
+  FilterPopup,
+  FilterPopupClose,
+  FilterPopupContainer,
+  FilterPopupContent,
+  FilterPopupHeader,
+  FilterPopupTitle,
+  FilterTabsRow,
+  SearchContainer,
+  ViewToggle,
+  ViewToggleButton,
+} from "./Documents.styles";
+import { DocumentsGridView } from "./DocumentsGridView";
+import { DocumentsListView } from "./DocumentsListView";
+import { DocumentsCompactView } from "./DocumentsCompactView";
 
 // ═══════════════════════════════════════════════════════════════════════════════
 // CONSTANTS
@@ -119,583 +128,6 @@ import {
 // cap (RELAY_CONNECTION_MAX_LIMIT = 100), which is the cost the slow render
 // was actually charged.
 const DOCUMENTS_PAGE_SIZE = 20;
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// STYLED COMPONENTS - Following CorpusListView/DiscoveryLanding patterns
-// ═══════════════════════════════════════════════════════════════════════════════
-
-const SearchContainer = styled.div`
-  margin-bottom: 16px;
-`;
-
-const FilterTabsRow = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 12px;
-`;
-
-const FilterButton = styled.button<{
-  $active?: boolean;
-  $hasFilters?: boolean;
-}>`
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  padding: 8px 12px;
-  background: ${(props) =>
-    props.$active ? OS_LEGAL_COLORS.surfaceLight : "white"};
-  border: 1px solid
-    ${(props) =>
-      props.$hasFilters ? OS_LEGAL_COLORS.accent : OS_LEGAL_COLORS.border};
-  border-radius: 8px;
-  color: ${(props) =>
-    props.$hasFilters ? OS_LEGAL_COLORS.accent : OS_LEGAL_COLORS.textSecondary};
-  font-size: 14px;
-  font-weight: 500;
-  cursor: pointer;
-  transition: all 0.15s;
-
-  &:hover {
-    background: ${OS_LEGAL_COLORS.surfaceHover};
-    border-color: ${(props) =>
-      props.$hasFilters ? OS_LEGAL_COLORS.accent : OS_LEGAL_COLORS.borderHover};
-  }
-
-  svg {
-    width: 16px;
-    height: 16px;
-  }
-`;
-
-const FilterBadge = styled.span`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  min-width: 18px;
-  height: 18px;
-  padding: 0 5px;
-  background: ${OS_LEGAL_COLORS.accent};
-  color: white;
-  font-size: 11px;
-  font-weight: 600;
-  border-radius: 9px;
-`;
-
-const FilterPopupContainer = styled.div`
-  position: relative;
-`;
-
-const FilterPopup = styled.div`
-  position: absolute;
-  top: calc(100% + 8px);
-  left: 0;
-  z-index: 50;
-  min-width: 320px;
-  padding: 16px;
-  background: white;
-  border: 1px solid ${OS_LEGAL_COLORS.border};
-  border-radius: 12px;
-  box-shadow: 0 10px 40px rgba(0, 0, 0, 0.12);
-
-  /* Override the harsh gradient labels from filter components */
-  .ui.label {
-    background: ${OS_LEGAL_COLORS.surfaceLight} !important;
-    color: ${OS_LEGAL_COLORS.textTertiary} !important;
-    box-shadow: none !important;
-    font-size: 0.6875rem !important;
-    font-weight: 600 !important;
-    letter-spacing: 0.05em !important;
-  }
-
-  @media (max-width: 640px) {
-    left: 50%;
-    transform: translateX(-50%);
-    min-width: calc(100vw - 48px);
-    max-width: 400px;
-  }
-`;
-
-const FilterPopupHeader = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 12px;
-  padding-bottom: 12px;
-  border-bottom: 1px solid ${OS_LEGAL_COLORS.border};
-`;
-
-const FilterPopupTitle = styled.span`
-  font-size: 14px;
-  font-weight: 600;
-  color: ${OS_LEGAL_COLORS.textPrimary};
-`;
-
-const FilterPopupClose = styled.button`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 28px;
-  height: 28px;
-  padding: 0;
-  background: transparent;
-  border: none;
-  border-radius: 6px;
-  color: ${OS_LEGAL_COLORS.textMuted};
-  cursor: pointer;
-
-  &:hover {
-    background: ${OS_LEGAL_COLORS.surfaceLight};
-    color: ${OS_LEGAL_COLORS.textTertiary};
-  }
-`;
-
-const FilterPopupContent = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-
-  /* Give each child descending z-index so earlier dropdowns appear above later ones */
-  & > *:nth-child(1) {
-    position: relative;
-    z-index: 30;
-  }
-  & > *:nth-child(2) {
-    position: relative;
-    z-index: 20;
-  }
-  & > *:nth-child(3) {
-    position: relative;
-    z-index: 10;
-  }
-  & > *:nth-child(4) {
-    position: relative;
-    z-index: 5;
-  }
-`;
-
-const ClearFiltersButton = styled.button`
-  margin-top: 8px;
-  padding: 8px 12px;
-  background: transparent;
-  border: 1px solid ${OS_LEGAL_COLORS.border};
-  border-radius: 6px;
-  color: ${OS_LEGAL_COLORS.textSecondary};
-  font-size: 13px;
-  font-weight: 500;
-  cursor: pointer;
-  transition: all 0.15s;
-
-  &:hover {
-    background: ${OS_LEGAL_COLORS.dangerSurface};
-    border-color: ${OS_LEGAL_COLORS.dangerBorder};
-    color: ${OS_LEGAL_COLORS.danger};
-  }
-`;
-
-const ActionButtons = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 12px;
-`;
-
-const ViewToggle = styled.div`
-  display: flex;
-  align-items: center;
-  background: white;
-  border: 1px solid ${OS_LEGAL_COLORS.border};
-  border-radius: 8px;
-  padding: 3px;
-`;
-
-const ViewToggleButton = styled.button<{ $active?: boolean }>`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 32px;
-  height: 32px;
-  padding: 0;
-  background: ${(props) =>
-    props.$active ? OS_LEGAL_COLORS.surfaceLight : "transparent"};
-  border: none;
-  border-radius: 6px;
-  color: ${(props) =>
-    props.$active ? OS_LEGAL_COLORS.textPrimary : OS_LEGAL_COLORS.textMuted};
-  cursor: pointer;
-  transition: all 0.15s;
-
-  &:hover {
-    color: ${OS_LEGAL_COLORS.textTertiary};
-  }
-`;
-
-const DocumentsListContainer = styled.section`
-  position: relative;
-  min-height: 200px;
-`;
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// DOCUMENT GRID STYLES
-// ═══════════════════════════════════════════════════════════════════════════════
-
-const DocumentsGrid = styled.div`
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(280px, 1fr));
-  gap: 20px;
-`;
-
-const DocumentCardWrapper = styled.div<{ $selected?: boolean }>`
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  background: white;
-  border: 1px solid
-    ${(props) =>
-      props.$selected ? OS_LEGAL_COLORS.accent : OS_LEGAL_COLORS.border};
-  border-radius: 12px;
-  overflow: hidden;
-  transition: all 0.15s;
-  cursor: pointer;
-  box-shadow: ${(props) =>
-    props.$selected ? `0 0 0 2px ${accentAlpha(0.2)}` : "none"};
-
-  &:hover {
-    border-color: ${(props) =>
-      props.$selected ? OS_LEGAL_COLORS.accent : OS_LEGAL_COLORS.borderHover};
-    box-shadow: ${(props) =>
-      props.$selected
-        ? `0 0 0 2px ${accentAlpha(0.2)}`
-        : "0 4px 6px rgba(15, 23, 42, 0.04)"};
-    transform: translateY(-2px);
-  }
-`;
-
-const CardCheckbox = styled.div<{ $visible?: boolean }>`
-  position: absolute;
-  top: 12px;
-  left: 12px;
-  z-index: 10;
-  opacity: ${(props) => (props.$visible ? 1 : 0)};
-  transition: opacity 0.15s;
-
-  ${DocumentCardWrapper}:hover & {
-    opacity: 1;
-  }
-`;
-
-const CardPreview = styled.div`
-  position: relative;
-  height: 160px;
-  background: linear-gradient(
-    135deg,
-    ${OS_LEGAL_COLORS.surfaceHover} 0%,
-    ${OS_LEGAL_COLORS.surfaceLight} 100%
-  );
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  overflow: hidden;
-`;
-
-const CardThumbnail = styled.img`
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  object-position: center top;
-`;
-
-const CardPreviewPlaceholder = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 8px;
-  padding: 20px;
-  color: ${OS_LEGAL_COLORS.textMuted};
-`;
-
-const PreviewLines = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 6px;
-  width: 80%;
-  max-width: 180px;
-`;
-
-const PreviewLine = styled.div<{ $width?: string }>`
-  height: 6px;
-  background: ${OS_LEGAL_COLORS.border};
-  border-radius: 3px;
-  width: ${(props) => props.$width || "100%"};
-`;
-
-const TypeBadge = styled.div`
-  position: absolute;
-  top: 12px;
-  right: 12px;
-`;
-
-const ProcessingOverlay = styled.div`
-  position: absolute;
-  inset: 0;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 12px;
-  background: rgba(248, 250, 252, 0.9);
-  backdrop-filter: blur(2px);
-`;
-
-const ProcessingText = styled.span`
-  font-size: 13px;
-  font-weight: 500;
-  color: ${OS_LEGAL_COLORS.textTertiary};
-`;
-
-const CardBody = styled.div`
-  padding: 16px;
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-`;
-
-const CardTitle = styled.h4`
-  font-size: 14px;
-  font-weight: 600;
-  color: ${OS_LEGAL_COLORS.textPrimary};
-  margin: 0;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-  line-height: 1.4;
-  word-break: break-word;
-`;
-
-const CardMeta = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  gap: 8px;
-  font-size: 12px;
-  color: ${OS_LEGAL_COLORS.textMuted};
-`;
-
-const MetaSeparator = styled.span`
-  width: 3px;
-  height: 3px;
-  border-radius: 50%;
-  background: ${OS_LEGAL_COLORS.textMuted};
-  opacity: 0.5;
-`;
-
-const CardFooter = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 12px 16px;
-  border-top: 1px solid ${OS_LEGAL_COLORS.border};
-  background: ${OS_LEGAL_COLORS.background};
-`;
-
-const CardUploader = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  font-size: 12px;
-  color: ${OS_LEGAL_COLORS.textTertiary};
-`;
-
-const CardMenuButton = styled.button`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 28px;
-  height: 28px;
-  background: transparent;
-  border: none;
-  border-radius: 6px;
-  color: ${OS_LEGAL_COLORS.textMuted};
-  cursor: pointer;
-  opacity: 0;
-  transition: all 0.15s;
-
-  ${DocumentCardWrapper}:hover & {
-    opacity: 1;
-  }
-
-  &:hover {
-    background: ${OS_LEGAL_COLORS.surfaceLight};
-    color: ${OS_LEGAL_COLORS.textTertiary};
-  }
-`;
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// DOCUMENT LIST STYLES
-// ═══════════════════════════════════════════════════════════════════════════════
-
-const DocumentsListView = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 2px;
-  background: white;
-  border: 1px solid ${OS_LEGAL_COLORS.border};
-  border-radius: 12px;
-  overflow: hidden;
-`;
-
-const ListHeader = styled.div`
-  display: grid;
-  grid-template-columns: 40px 1fr 100px 100px 120px 150px 48px;
-  gap: 16px;
-  padding: 12px 16px;
-  background: ${OS_LEGAL_COLORS.background};
-  border-bottom: 1px solid ${OS_LEGAL_COLORS.border};
-  font-size: 11px;
-  font-weight: 600;
-  text-transform: uppercase;
-  letter-spacing: 0.05em;
-  color: ${OS_LEGAL_COLORS.textMuted};
-
-  @media (max-width: 768px) {
-    grid-template-columns: 32px 1fr 80px 48px;
-
-    & > :nth-child(3),
-    & > :nth-child(5),
-    & > :nth-child(6) {
-      display: none;
-    }
-  }
-`;
-
-const ListItem = styled.div<{ $selected?: boolean }>`
-  display: grid;
-  grid-template-columns: 40px 1fr 100px 100px 120px 150px 48px;
-  gap: 16px;
-  padding: 12px 16px;
-  align-items: center;
-  cursor: pointer;
-  transition: background 0.1s;
-  background: ${(props) =>
-    props.$selected ? accentAlpha(0.04) : "transparent"};
-
-  &:hover {
-    background: ${(props) =>
-      props.$selected ? accentAlpha(0.06) : OS_LEGAL_COLORS.surfaceHover};
-  }
-
-  &:not(:last-child) {
-    border-bottom: 1px solid ${OS_LEGAL_COLORS.border};
-  }
-
-  @media (max-width: 768px) {
-    grid-template-columns: 32px 1fr 80px 48px;
-
-    & > :nth-child(3),
-    & > :nth-child(5),
-    & > :nth-child(6) {
-      display: none;
-    }
-  }
-`;
-
-const ListItemIcon = styled.div`
-  width: 32px;
-  height: 32px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: ${OS_LEGAL_COLORS.textSecondary};
-`;
-
-const ListItemName = styled.span`
-  font-size: 14px;
-  font-weight: 500;
-  color: ${OS_LEGAL_COLORS.textPrimary};
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-`;
-
-const ListItemType = styled.span`
-  font-size: 12px;
-  text-transform: uppercase;
-  color: ${OS_LEGAL_COLORS.textMuted};
-`;
-
-const ListItemSize = styled.span`
-  font-size: 13px;
-  color: ${OS_LEGAL_COLORS.textTertiary};
-`;
-
-const ListItemUploader = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  font-size: 13px;
-  color: ${OS_LEGAL_COLORS.textTertiary};
-`;
-
-const ListItemActions = styled.div`
-  display: flex;
-  justify-content: flex-end;
-  opacity: 0;
-  transition: opacity 0.1s;
-
-  ${ListItem}:hover & {
-    opacity: 1;
-  }
-`;
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// COMPACT VIEW STYLES
-// ═══════════════════════════════════════════════════════════════════════════════
-
-const DocumentsCompactView = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 2px;
-  background: white;
-  border: 1px solid ${OS_LEGAL_COLORS.border};
-  border-radius: 12px;
-  overflow: hidden;
-`;
-
-const CompactItem = styled.div<{ $selected?: boolean }>`
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  padding: 8px 16px;
-  cursor: pointer;
-  transition: background 0.1s;
-  background: ${(props) =>
-    props.$selected ? accentAlpha(0.04) : "transparent"};
-
-  &:hover {
-    background: ${(props) =>
-      props.$selected ? accentAlpha(0.06) : OS_LEGAL_COLORS.surfaceHover};
-  }
-
-  &:not(:last-child) {
-    border-bottom: 1px solid ${OS_LEGAL_COLORS.border};
-  }
-`;
-
-const CompactItemName = styled.span`
-  flex: 1;
-  font-size: 13px;
-  font-weight: 500;
-  color: ${OS_LEGAL_COLORS.textPrimary};
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-`;
-
-const CompactItemMeta = styled.span`
-  font-size: 12px;
-  color: ${OS_LEGAL_COLORS.textMuted};
-  flex-shrink: 0;
-`;
 
 // ═══════════════════════════════════════════════════════════════════════════════
 // ICONS
@@ -724,34 +156,6 @@ const DocumentIcon = () => (
     />
   </svg>
 );
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// HELPER FUNCTIONS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-function getDocumentType(doc: DocumentType): string {
-  // Use fileType field directly if available (preferred)
-  if (doc.fileType) {
-    const ft = doc.fileType.toLowerCase();
-    if (ft === "pdf") return "PDF";
-    if (ft === "docx" || ft === "doc") return "DOCX";
-    if (ft === "txt") return "TXT";
-    return ft.toUpperCase();
-  }
-  // Fallback: parse from title if it has a valid extension
-  const fileName = doc.title || "";
-  const parts = fileName.split(".");
-  // Only use extension if there are multiple parts (filename.ext)
-  if (parts.length > 1) {
-    const ext = parts.pop()?.toLowerCase();
-    if (ext === "pdf") return "PDF";
-    if (ext === "docx" || ext === "doc") return "DOCX";
-    if (ext === "txt") return "TXT";
-    if (ext) return ext.toUpperCase();
-  }
-  // Default to PDF if no extension found
-  return "PDF";
-}
 
 // ═══════════════════════════════════════════════════════════════════════════════
 // MAIN COMPONENT
@@ -1094,6 +498,10 @@ export const Documents = () => {
     setShowFilterPopup(false);
   };
 
+  const allSelected =
+    selected_document_ids.length === filteredDocuments.length &&
+    filteredDocuments.length > 0;
+
   return (
     <PageContainer>
       <ContentContainer>
@@ -1235,7 +643,7 @@ export const Documents = () => {
         </StatsContainer>
 
         {/* Documents Section */}
-        <DocumentsListContainer>
+        <DocumentsSection>
           {/* Cover the grid only on the initial load — fetchMore keeps existing rows visible. */}
           <LoadingOverlay
             active={documents_loading && filteredDocuments.length === 0}
@@ -1329,256 +737,38 @@ export const Documents = () => {
           {filteredDocuments.length > 0 ? (
             <>
               {viewMode === VIEW_MODES.GRID && (
-                <DocumentsGrid>
-                  {filteredDocuments.map((doc) => (
-                    <DocumentCardWrapper
-                      key={doc.id}
-                      role="button"
-                      tabIndex={0}
-                      data-testid="document-card"
-                      data-processing={String(Boolean(doc.backendLock))}
-                      aria-label={`Open document ${doc.title || "Untitled"}`}
-                      $selected={selected_document_ids.includes(doc.id)}
-                      onClick={() => handleDocumentClick(doc)}
-                      onContextMenu={(e) => handleContextMenu(e, doc)}
-                      onKeyDown={(e) => {
-                        if (e.key === "Enter" || e.key === " ") {
-                          e.preventDefault();
-                          handleDocumentClick(doc);
-                        }
-                      }}
-                    >
-                      <CardCheckbox
-                        $visible={selected_document_ids.includes(doc.id)}
-                        onClick={(e) => e.stopPropagation()}
-                      >
-                        <input
-                          type="checkbox"
-                          aria-label={`Select ${doc.title || "Untitled"}`}
-                          checked={selected_document_ids.includes(doc.id)}
-                          onChange={() => handleSelect(doc.id)}
-                        />
-                      </CardCheckbox>
-
-                      <CardPreview>
-                        {doc.icon ? (
-                          <CardThumbnail
-                            src={doc.icon}
-                            alt={doc.title || "Document"}
-                          />
-                        ) : (
-                          <CardPreviewPlaceholder>
-                            <FileText size={48} />
-                            <PreviewLines>
-                              <PreviewLine />
-                              <PreviewLine $width="85%" />
-                              <PreviewLine $width="90%" />
-                              <PreviewLine $width="70%" />
-                            </PreviewLines>
-                          </CardPreviewPlaceholder>
-                        )}
-
-                        <TypeBadge>
-                          <Chip size="sm" variant="filled" color="default">
-                            {getDocumentType(doc)}
-                          </Chip>
-                        </TypeBadge>
-
-                        {doc.backendLock && (
-                          <ProcessingOverlay>
-                            <Loader2
-                              size={24}
-                              className="animate-spin"
-                              style={{ animation: "spin 1s linear infinite" }}
-                            />
-                            <ProcessingText>Processing...</ProcessingText>
-                          </ProcessingOverlay>
-                        )}
-                      </CardPreview>
-
-                      <CardBody>
-                        <CardTitle title={doc.title || "Untitled"}>
-                          {doc.title || "Untitled"}
-                        </CardTitle>
-                        <CardMeta>
-                          {doc.pageCount ? (
-                            <span>{doc.pageCount} pages</span>
-                          ) : (
-                            <span>Document</span>
-                          )}
-                        </CardMeta>
-                      </CardBody>
-
-                      <CardFooter>
-                        <CardUploader>
-                          <Avatar
-                            fallback={getCreatorInitials(doc.creator)}
-                            size="xs"
-                          />
-                          <span>{formatRelativeTime(doc.created)}</span>
-                        </CardUploader>
-                        <CardMenuButton
-                          aria-label="Open menu"
-                          aria-haspopup="menu"
-                          aria-expanded={contextMenu?.document.id === doc.id}
-                          onClick={(e) => {
-                            e.stopPropagation();
-                            handleContextMenu(e, doc);
-                          }}
-                        >
-                          <MoreVertical size={16} />
-                        </CardMenuButton>
-                      </CardFooter>
-                    </DocumentCardWrapper>
-                  ))}
-                </DocumentsGrid>
+                <DocumentsGridView
+                  documents={filteredDocuments}
+                  selectedIds={selected_document_ids}
+                  activeContextMenuDocId={contextMenu?.document.id}
+                  onDocumentClick={handleDocumentClick}
+                  onSelect={handleSelect}
+                  onContextMenu={handleContextMenu}
+                />
               )}
 
               {viewMode === VIEW_MODES.LIST && (
-                <DocumentsListView role="table" aria-label="Documents list">
-                  <ListHeader role="rowgroup">
-                    <input
-                      type="checkbox"
-                      aria-label="Select all documents"
-                      checked={
-                        selected_document_ids.length ===
-                          filteredDocuments.length &&
-                        filteredDocuments.length > 0
-                      }
-                      onChange={handleSelectAll}
-                    />
-                    <span>Name</span>
-                    <span>Type</span>
-                    <span>Pages</span>
-                    <span>Status</span>
-                    <span>Uploaded</span>
-                    <span></span>
-                  </ListHeader>
-                  {filteredDocuments.map((doc) => (
-                    <ListItem
-                      key={doc.id}
-                      role="row"
-                      tabIndex={0}
-                      data-testid="document-card"
-                      data-processing={String(Boolean(doc.backendLock))}
-                      aria-label={`Open document ${doc.title || "Untitled"}`}
-                      $selected={selected_document_ids.includes(doc.id)}
-                      onClick={() => handleDocumentClick(doc)}
-                      onContextMenu={(e) => handleContextMenu(e, doc)}
-                      onKeyDown={(e) => {
-                        if (e.key === "Enter" || e.key === " ") {
-                          e.preventDefault();
-                          handleDocumentClick(doc);
-                        }
-                      }}
-                    >
-                      <div onClick={(e) => e.stopPropagation()}>
-                        <input
-                          type="checkbox"
-                          aria-label={`Select ${doc.title || "Untitled"}`}
-                          checked={selected_document_ids.includes(doc.id)}
-                          onChange={() => handleSelect(doc.id)}
-                        />
-                      </div>
-                      <ListItemName title={doc.title || "Untitled"}>
-                        {doc.title || "Untitled"}
-                      </ListItemName>
-                      <ListItemType>{getDocumentType(doc)}</ListItemType>
-                      <ListItemSize>
-                        {doc.pageCount ? `${doc.pageCount} pages` : ""}
-                      </ListItemSize>
-                      <div>
-                        <Chip
-                          size="sm"
-                          variant="soft"
-                          color={doc.backendLock ? "warning" : "success"}
-                        >
-                          {doc.backendLock ? "Processing" : "Processed"}
-                        </Chip>
-                      </div>
-                      <ListItemUploader>
-                        <Avatar
-                          fallback={getCreatorInitials(doc.creator)}
-                          size="xs"
-                        />
-                        <span>{formatRelativeTime(doc.created)}</span>
-                      </ListItemUploader>
-                      <ListItemActions>
-                        <CardMenuButton
-                          aria-label="Open menu"
-                          aria-haspopup="menu"
-                          aria-expanded={contextMenu?.document.id === doc.id}
-                          onClick={(e) => {
-                            e.stopPropagation();
-                            handleContextMenu(e, doc);
-                          }}
-                        >
-                          <MoreVertical size={16} />
-                        </CardMenuButton>
-                      </ListItemActions>
-                    </ListItem>
-                  ))}
-                </DocumentsListView>
+                <DocumentsListView
+                  documents={filteredDocuments}
+                  selectedIds={selected_document_ids}
+                  activeContextMenuDocId={contextMenu?.document.id}
+                  allSelected={allSelected}
+                  onDocumentClick={handleDocumentClick}
+                  onSelect={handleSelect}
+                  onSelectAll={handleSelectAll}
+                  onContextMenu={handleContextMenu}
+                />
               )}
 
               {viewMode === VIEW_MODES.COMPACT && (
-                <DocumentsCompactView>
-                  {filteredDocuments.map((doc) => (
-                    <CompactItem
-                      key={doc.id}
-                      role="listitem"
-                      tabIndex={0}
-                      data-testid="document-card"
-                      data-processing={String(Boolean(doc.backendLock))}
-                      aria-label={`Open document ${doc.title || "Untitled"}`}
-                      $selected={selected_document_ids.includes(doc.id)}
-                      onClick={() => handleDocumentClick(doc)}
-                      onContextMenu={(e) => handleContextMenu(e, doc)}
-                      onKeyDown={(e) => {
-                        if (e.key === "Enter" || e.key === " ") {
-                          e.preventDefault();
-                          handleDocumentClick(doc);
-                        }
-                      }}
-                    >
-                      <div onClick={(e) => e.stopPropagation()}>
-                        <input
-                          type="checkbox"
-                          aria-label={`Select ${doc.title || "Untitled"}`}
-                          checked={selected_document_ids.includes(doc.id)}
-                          onChange={() => handleSelect(doc.id)}
-                        />
-                      </div>
-                      <ListItemIcon>
-                        <FileText size={20} />
-                      </ListItemIcon>
-                      <CompactItemName title={doc.title || "Untitled"}>
-                        {doc.title || "Untitled"}
-                      </CompactItemName>
-                      <CompactItemMeta>
-                        {doc.pageCount ? `${doc.pageCount} pages` : ""}
-                      </CompactItemMeta>
-                      <Chip
-                        size="sm"
-                        variant="soft"
-                        color={doc.backendLock ? "warning" : "success"}
-                      >
-                        {doc.backendLock ? "Processing" : "Processed"}
-                      </Chip>
-                      <CardMenuButton
-                        aria-label="Open menu"
-                        aria-haspopup="menu"
-                        aria-expanded={contextMenu?.document.id === doc.id}
-                        onClick={(e) => {
-                          e.stopPropagation();
-                          handleContextMenu(e, doc);
-                        }}
-                      >
-                        <MoreVertical size={16} />
-                      </CardMenuButton>
-                    </CompactItem>
-                  ))}
-                </DocumentsCompactView>
+                <DocumentsCompactView
+                  documents={filteredDocuments}
+                  selectedIds={selected_document_ids}
+                  activeContextMenuDocId={contextMenu?.document.id}
+                  onDocumentClick={handleDocumentClick}
+                  onSelect={handleSelect}
+                  onContextMenu={handleContextMenu}
+                />
               )}
 
               <FetchMoreOnVisible fetchNextPage={handleFetchMore} />
@@ -1640,7 +830,7 @@ export const Documents = () => {
               />
             </EmptyStateWrapper>
           ) : null}
-        </DocumentsListContainer>
+        </DocumentsSection>
 
         {/* Context Menu */}
         {contextMenu && (
